Image fetcher auth
StackMirror can have an auth tuple set on it. When ``mirror`` is set on
the ``StackMirror``, that auth tuple is added to the requests session.

The ImageFetcher can also have an auth tuple set on it, which it falls
back on if the selected mirror does not have one of its own.